requirements.txt
pyaudio
google-generativeai
boto3 
pygame
python-dotenv
typing
threading
pathlib